requirements_panda.txt
absl-py==2.1.0
alembic==1.13.1
asttokens==2.4.1
async-timeout==4.0.3
backcall==0.2.0
bitmath==1.3.3.1
brotlipy==0.7.0
cachetools==5.3.3
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet==4.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==7.1.2
cloudpickle==3.0.0
colorlog==6.8.2
contourpy==1.1.1
cryptography @ file:///tmp/build/80754af9/cryptography_1652083738073/work
cycler==0.12.1
DataProperty==1.0.1
decorator==5.1.1
deprecation==2.1.0
DI-engine==0.4.9
DI-toolkit==0.2.1
DI-treetensor==0.4.1
dill==0.3.8
docker-pycreds==0.4.0
easydict==1.9
enum-tools==0.12.0
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.14.0
Flask==1.1.4
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.63.0
gym==0.25.1
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.11.0
hbutils==0.9.3
hickle==5.0.3
huggingface-hub==0.23.0
huggingface-sb3==3.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==2.11.3
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.41.1
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.7.5
matplotlib-inline==0.1.7
mbstrdecoder==1.1.3
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpire==2.10.2
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpngw==0.1.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
optuna==3.6.1
packaging==24.0
panda-gym==3.0.7
pandas==2.0.3
parso==0.8.4
pathvalidate==3.2.0
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybullet==3.2.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.18.0
pynng==0.8.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytablewriter==1.2.0
python-dateutil==2.9.0.post0
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==2.0.0
responses==0.12.1
rich==13.7.1
rl-zoo3==2.2.1
rsa==4.9
sb3_contrib==2.3.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.1.1
setproctitle==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.30
stable_baselines3==2.3.2
stack-data==0.6.3
statsmodels==0.14.0
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.3.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.66.4
traitlets==5.14.3
treevalue==1.4.12
triton==2.3.0
trueskill==0.4.5
typepy==1.3.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
URLObject==2.4.3
wandb==0.17.0
wasabi==1.1.2
wcwidth==0.2.13
Werkzeug==1.0.1
yapf==0.29.0
yattag==1.15.2
zipp==3.18.1
